[feat] add endpoint for external extensions to automatically register as an integration
External extensions can now send a request to Blue Blocker to be registered as a new integration. New integrations registered in this way are set to the disabled state by default, and the user of Blue Block still has to manually set the permissions of the new integration in order for the integration functionality to be enabled. This feature just makes it so that extension users do not have to find the extension ID of an external extension to register an integration.